+++ packages/inngest/src/components/connect/index.ts
@@ -1,24 +1,8 @@
-import { envKeys, headerKeys, queryKeys } from "../../helpers/consts.ts";
-import { allProcessEnv, getEnvironmentName } from "../../helpers/env.ts";
-import { parseFnData } from "../../helpers/functions.ts";
-import { hashSigningKey } from "../../helpers/strings.ts";
-import {
-  type GatewayExecutorRequestData,
-  SDKResponse,
-  SDKResponseStatus,
-} from "../../proto/src/components/connect/protobuf/connect.ts";
-import type { Capabilities, FunctionConfig } from "../../types.ts";
-import { version } from "../../version.ts";
-import { PREFERRED_ASYNC_EXECUTION_VERSION } from "../execution/InngestExecution.ts";
+import { envKeys } from "../../helpers/consts.ts";
+import { allProcessEnv } from "../../helpers/env.ts";
 import { type Inngest, internalLoggerSymbol } from "../Inngest.ts";
-import { InngestCommHandler } from "../InngestCommHandler.ts";
-import type { InngestFunction } from "../InngestFunction.ts";
-import {
-  type ConnectionEstablishData,
-  type ConnectionStrategy,
-  createStrategy,
-  type RequestHandler,
-} from "./strategies/index.ts";
+import { prepareConnectionConfig } from "./config.ts";
+import { type ConnectionStrategy, createStrategy } from "./strategies/index.ts";
 import {
   type ConnectApp,
   type ConnectHandlerOptions,
@@ -26,16 +10,6 @@ import {
   DEFAULT_SHUTDOWN_SIGNALS,
   type WorkerConnection,
 } from "./types.ts";
-import { parseTraceCtx } from "./util.ts";
-
-const InngestBranchEnvironmentSigningKeyPrefix = "signkey-branch-";
-
-type ConnectCommHandler = InngestCommHandler<
-  [GatewayExecutorRequestData],
-  SDKResponse,
-  // biome-ignore lint/suspicious/noExplicitAny: intentional
-  any
->;
 
 /**
  * WebSocket worker connection that implements the WorkerConnection interface.
@@ -70,35 +44,6 @@ class WebSocketWorkerConnection implements WorkerConnection {
     }
 
     this.options = this.applyDefaults(options);
-  }
-
-  private get functions(): Record<
-    string,
-    {
-      client: Inngest.Like;
-      functions: InngestFunction.Any[];
-    }
-  > {
-    const functions: Record<
-      string,
-      {
-        client: Inngest.Like;
-        functions: InngestFunction.Any[];
-      }
-    > = {};
-    for (const app of this.options.apps) {
-      const client = app.client as Inngest.Any;
-
-      if (functions[client.id]) {
-        throw new Error(`Duplicate app id: ${client.id}`);
-      }
-
-      functions[client.id] = {
-        client: app.client,
-        functions: (app.functions as InngestFunction.Any[]) ?? client.funcs,
-      };
-    }
-    return functions;
   }
 
   private applyDefaults(opts: ConnectHandlerOptions): ConnectHandlerOptions {
@@ -171,188 +116,13 @@ class WebSocketWorkerConnection implements WorkerConnection {
       "Establishing connection",
     );
 
-    const envName = this.inngest.env ?? getEnvironmentName();
-
-    const hashedSigningKey = this.inngest.signingKey
-      ? hashSigningKey(this.inngest.signingKey)
-      : undefined;
-
-    if (
-      this.inngest.signingKey &&
-      this.inngest.signingKey.startsWith(
-        InngestBranchEnvironmentSigningKeyPrefix,
-      ) &&
-      !envName
-    ) {
-      throw new Error(
-        "Environment is required when using branch environment signing keys",
-      );
-    }
-
-    const hashedFallbackKey = this.inngest.signingKeyFallback
-      ? hashSigningKey(this.inngest.signingKeyFallback)
-      : undefined;
-
-    // Build capabilities
-    const capabilities: Capabilities = {
-      trust_probe: "v1",
-      connect: "v1",
-    };
-
-    // Build function configs
-    const functionConfigs: Record<
-      string,
-      {
-        client: Inngest.Like;
-        functions: FunctionConfig[];
-      }
-    > = {};
-    for (const [appId, { client, functions }] of Object.entries(
-      this.functions,
-    )) {
-      functionConfigs[appId] = {
-        client: client,
-        functions: functions.flatMap((f) =>
-          f["getConfig"]({
-            baseUrl: new URL("wss://connect"),
-            appPrefix: (client as Inngest.Any).id,
-            isConnect: true,
-          }),
-        ),
-      };
-    }
-
-    this.inngest[internalLoggerSymbol].debug(
-      {
-        functionSlugs: Object.entries(functionConfigs).map(
-          ([appId, { functions }]) => {
-            return JSON.stringify({
-              appId,
-              functions: functions.map((f) => ({
-                id: f.id,
-                stepUrls: Object.values(f.steps).map((s) => s.runtime["url"]),
-              })),
-            });
-          },
-        ),
-      },
-      "Prepared sync data",
-    );
-
-    // Build connection establish data
-    const connectionData: ConnectionEstablishData = {
-      manualReadinessAck: false,
-      marshaledCapabilities: JSON.stringify(capabilities),
-      apps: Object.entries(functionConfigs).map(
-        ([appId, { client, functions }]) => ({
-          appName: appId,
-          appVersion: (client as Inngest.Any).appVersion,
-          functions: new TextEncoder().encode(JSON.stringify(functions)),
-        }),
-      ),
-    };
-
-    // Build request handlers
-    const requestHandlers: Record<string, RequestHandler> = {};
-    for (const [appId, { client, functions }] of Object.entries(
-      this.functions,
-    )) {
-      const inngestCommHandler: ConnectCommHandler = new InngestCommHandler({
-        client: client,
-        functions: functions,
-        frameworkName: "connect",
-        skipSignatureValidation: true,
-        handler: (msg: GatewayExecutorRequestData) => {
-          const asString = new TextDecoder().decode(msg.requestPayload);
-          const parsed = parseFnData(
-            JSON.parse(asString),
-            undefined,
-            this.inngest[internalLoggerSymbol],
-          );
-
-          const userTraceCtx = parseTraceCtx(msg.userTraceCtx);
-
-          return {
-            body() {
-              return parsed;
-            },
-            method() {
-              return "POST";
-            },
-            headers(key) {
-              switch (key) {
-                case headerKeys.ContentLength.toString():
-                  return asString.length.toString();
-                case headerKeys.InngestExpectedServerKind.toString():
-                  return "connect";
-                case headerKeys.RequestVersion.toString():
-                  return parsed.version.toString();
-                case headerKeys.Signature.toString():
-                  return null;
-                case headerKeys.TraceParent.toString():
-                  return userTraceCtx?.traceParent ?? null;
-                case headerKeys.TraceState.toString():
-                  return userTraceCtx?.traceState ?? null;
-                default:
-                  return null;
-              }
-            },
-            transformResponse({ body, headers, status }) {
-              let sdkResponseStatus: SDKResponseStatus = SDKResponseStatus.DONE;
-              switch (status) {
-                case 200:
-                  sdkResponseStatus = SDKResponseStatus.DONE;
-                  break;
-                case 206:
-                  sdkResponseStatus = SDKResponseStatus.NOT_COMPLETED;
-                  break;
-                case 500:
-                  sdkResponseStatus = SDKResponseStatus.ERROR;
-                  break;
-              }
-
-              return SDKResponse.create({
-                requestId: msg.requestId,
-                accountId: msg.accountId,
-                envId: msg.envId,
-                appId: msg.appId,
-                status: sdkResponseStatus,
-                body: new TextEncoder().encode(body),
-                noRetry: headers[headerKeys.NoRetry] === "true",
-                retryAfter: headers[headerKeys.RetryAfter],
-                sdkVersion: `inngest-js:v${version}`,
-                requestVersion: parseInt(
-                  headers[headerKeys.RequestVersion] ??
-                    PREFERRED_ASYNC_EXECUTION_VERSION.toString(),
-                  10,
-                ),
-                systemTraceCtx: msg.systemTraceCtx,
-                userTraceCtx: msg.userTraceCtx,
-                runId: msg.runId,
-              });
-            },
-            url() {
-              const baseUrl = new URL("http://connect.inngest.com");
-
-              baseUrl.searchParams.set(queryKeys.FnId, msg.functionSlug);
-
-              if (msg.stepId) {
-                baseUrl.searchParams.set(queryKeys.StepId, msg.stepId);
-              }
-
-              return baseUrl;
-            },
-          };
-        },
-      });
-
-      if (!inngestCommHandler.checkModeConfiguration()) {
-        throw new Error("Signing key is required");
-      }
-
-      const requestHandler = inngestCommHandler.createHandler();
-      requestHandlers[appId] = requestHandler;
-    }
+    const {
+      hashedSigningKey,
+      hashedFallbackKey,
+      envName,
+      connectionData,
+      requestHandlers,
+    } = prepareConnectionConfig(this.options.apps, this.inngest);
 
     // Create and initialize the strategy
     this.strategy = await createStrategy(
