+++ packages/inngest/src/components/connect/strategies/core/index.ts
@@ -4,11 +4,16 @@ export {
   type ConnectionCoreCallbacks,
   type ConnectionCoreConfig,
 } from "./connection.ts";
+export { establishConnection } from "./handshake.ts";
+export { HeartbeatManager } from "./heartbeat.ts";
+export { RequestProcessor } from "./requestProcessor.ts";
 
 export type {
+  ConnectionAccessor,
   ConnectionEstablishData,
   ConnectionStrategy,
   RequestHandler,
   StrategyConfig,
   StrategyEvents,
+  WakeSignal,
 } from "./types.ts";
