+++ packages/inngest/src/components/connect/strategies/core/types.ts
@@ -1,3 +1,4 @@
+import type { WaitGroup } from "@jpwilliams/waitgroup";
 import type { Mode } from "../../../../helpers/env.ts";
 import type { Logger } from "../../../../middleware/logger.ts";
 import type {
@@ -6,6 +7,7 @@ import type {
 } from "../../../../proto/src/components/connect/protobuf/connect.ts";
 import type { Inngest } from "../../../Inngest.ts";
 import type { ConnectHandlerOptions, ConnectionState } from "../../types.ts";
+import type { Connection } from "./connection.ts";
 
 /**
  * A request handler that processes executor requests and returns SDK responses.
@@ -133,4 +135,26 @@ export interface ConnectionStrategy {
    * A promise that resolves when the connection is fully closed.
    */
   readonly closed: Promise<void>;
+}
+
+/**
+ * Narrow view of ConnectionCore state, passed to sub-modules so they can
+ * read (but not mutate) the connection state they need.
+ */
+export interface ConnectionAccessor {
+  readonly activeConnection: Connection | undefined;
+  readonly drainingConnection: Connection | undefined;
+  readonly shutdownRequested: boolean;
+  readonly inProgressRequests: {
+    wg: WaitGroup;
+    requestLeases: Record<string, string>;
+  };
+  readonly appIds: string[];
+}
+
+/**
+ * Wake signal interface for sub-modules to wake the reconcile loop.
+ */
+export interface WakeSignal {
+  wake(): void;
 }
